add example and classifier option
Some cleanup -- separate the act of classifying from the act of training, so you don't need to load
the base data to reuse an old model.

There are still some awkward bits here around how we deal with the neural net, but it is nice to have
clean, usable code to experiment around.